BUG: Pattern native libraries for Linux used the wrong namespace.
As the namespace used to build the library did not match the one used to when accessing via the interop layer, a missing symbol exception is thrown.
Libraries are now built using the correct namespace so the library exports the same symbols as those called by the interop layer.

Also marked PreBuild.sh as executable so it does not need to be modified in the build stage.