BUG: Fix a bug that causes an Exception when using the cloud engine and accessing properties that have 'HasValue = false'.
The 'noValueReasons' in the JSON have the complete name of the property (e.g. device.ismobile)
When the DeviceDataCloud instance is created, these are placed in a dictionary using only the second part of the name as the key (e.g. ismobile)

However, when retrieving the reason for a null value in the 'SetValue' method, it looks in the dictionary using the 2 part name. Consequently, it fails to find anything.

This has now been corrected.