Merged PR 7529: BUILD: Native DLLs in Framwork are now copied to the runtimes directory, not the root of the bin directory, and the default for AnyCPU is now x64 instead of x86.
The behaviour of the NuGet package for .NET Framework now uses `runtimes/win-x64/native` and `runtimes/win-x86/native` for the native DLLs. These are automatically loaded by the runtime for x64 and x86 builds respectively. For AnyCPU builds, the x64 native DLL is copied to the root of the output. This is more likely to be correct than x86.
#4935

Related work items: #4935